Add a fix for proton games
Some games need the surround sound sink to be the default when starting in order to negotiate channel count. After starting, these apps will never re-negitiate.
This changeset should modify the SteamDeck audio config enough and provide the plugin with a config option to set the Virtual Surround Sound sink as the default for all audio. Chances are, this is what most people want, so this will be the default.